requirements.txt
PyYAML==6.0.1
requests==2.28.1
dnspython==2.2.1
argparse==1.4.0